12-requirements.txt
Pillow==10.3.0
pytorch_lightning==1.9.5
torch==1.10.2
torchmetrics==1.1.2
torchvision==0.11.3
